+++ packages/sitecore-jss-nextjs/src/index.ts
@@ -38,9 +38,7 @@ export {
   ComponentParams,
   getContentStylesheetLink,
   EditMode,
-  RenderingType,
 } from '@sitecore-jss/sitecore-jss/layout';
-export { RestComponentLayoutService } from '@sitecore-jss/sitecore-jss/editing';
 export { mediaApi } from '@sitecore-jss/sitecore-jss/media';
 export {
   trackingApi,
@@ -141,11 +139,9 @@ export {
   BYOCComponentParams,
   BYOCComponent,
   BYOCComponentProps,
-  getDesignLibraryStylesheetLinks,
   File,
   FileField,
   RichTextField,
-  DesignLibrary,
   DefaultEmptyFieldEditingComponentImage,
   DefaultEmptyFieldEditingComponentText,
   VisitorIdentification,